+++ src/clearcontrol/microscope/lightsheet/calibrator/modules/CalibrationA.java
@@ -9,9 +9,6 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
-import net.imglib2.img.basictypeaccess.offheap.ShortOffHeapAccess;
-import net.imglib2.img.planar.OffHeapPlanarImg;
-import net.imglib2.type.numeric.integer.UnsignedShortType;
 import clearcontrol.core.math.argmax.ArgMaxFinder1DInterface;
 import clearcontrol.core.math.argmax.Fitting1D;
 import clearcontrol.core.math.argmax.SmartArgMaxFinder;
@@ -26,6 +23,9 @@ import clearcontrol.microscope.lightsheet.component.detection.DetectionArmInterf
 import clearcontrol.microscope.lightsheet.component.lightsheet.LightSheetInterface;
 import clearcontrol.stack.StackInterface;
 import gnu.trove.list.array.TDoubleArrayList;
+import net.imglib2.img.basictypeaccess.offheap.ShortOffHeapAccess;
+import net.imglib2.img.planar.OffHeapPlanarImg;
+import net.imglib2.type.numeric.integer.UnsignedShortType;
 
 public class CalibrationA
 {
@@ -78,8 +78,8 @@ public class CalibrationA
     double lMinA = -25;
     double lMaxA = 25;
 
-    double lMinY = lLightSheet.getYVariable().getMin();
-    double lMaxY = lLightSheet.getYVariable().getMax();
+    double lMinY = lLightSheet.getYVariable().getMin().doubleValue();
+    double lMaxY = lLightSheet.getYVariable().getMax().doubleValue();
 
     double[] angles = new double[mNumberOfDetectionArmDevices];
     int lCount = 0;
